Adjust panel position
Adjust panel position and fix Maxwell problems

The panel was originally sitting too high so it
was moved down. Maxwell was stationary in World
Space rather than being parented to the player, so
now he is parented to the Goto Recall panel. He
also did not disappear after getting an answer
correct/incorrect, so a timer was added to make
him hide after a few seconds.